chore(deploy): get and commit locales before tagging release
Previously, they were fetched and applied before building as part of
release, but not committed, so the release may have shipped a state of
repository which didn't correspond to the tag. This change makes sure
locales are fetched only as part of ./script/deploy-production.sh, and
any changes are committed before tagging.