feat: improved updater (#287)
* chore(updater): set downloads directory to Updates in userData

* chore(updater): remove old unnecessary code

We don't use Squirrel.Windows for a long time.

* feat: handle update failures

When starting, Peerio checks if the last update failed (this is designed
to happen when starting a new version right after the update), and if
so, it will show the error to the user and ask them to try the update
again, download the new version manually, or continue use the current
(older) version.

This also changes the "update is ready to install" dialog from alert()
to PeerUI Dialog and introduces another dialog to indicate that the
update is in progress.

Languages are now loaded before mounting the root component to avoid
race condition when something may have been showed before translations
are initialized.

* feat(updater): force check and install if client is deprecated

* feat: install updates on quit, ask to update every 12 hours

When the update is downloaded and it's not mandatory, it's automatically
scheduled to be installed when the user quits the app. If the user
doesn't quit within 12 hours, they are asked to install update (via a
dialog), with an option to postpone ("Later" button). The user will be
asked again in 12 hours, up to 2 times. The third time the dialog
appears, there's no "Later" option and the user will have to update.

If the update is mandatory, the user is asked to install it immediately
without an option to postpone.

* fix(updater): use isMandatorySince()

* fix(updater): pass update-downloaded event args correctly

* fix(updater): import app from remote

* fix(updater): pass mandatory status with update-downloaded event

* fix(updater): swap Later/Update button order

* chore(deps): bump desktop-release-builder version

* fix(updater): scheduleInstallOnQuit even for mandatory

* fix(updater): do not check for updates if last update failed

* chore(updater): lint

* chore: test release

* chore: test release

* chore: test release

* chore(deps): update release builder version

* chore: mandatory option for deploy-production.sh

Optional update:

	./deploy-production.sh

Mandatory update:

	./deploy-production mandatory

* chore: set lastMandatoryVersion to 2.114.3

* fix(updater): race condition when server deprecates client

Client checks for update and downloads it.
Client tries to login, server says it's deprecated.
Client calls quitAndRetryInstall and the updater tries download it
again, causing the error, causing the restart of the app.

Protect against calling quitAndRetryInstall and also make update
mandatory regardless of what's specified in the manifest if the server
reports that the client is deprecated.

* chore: bump lastMandatoryUpdateVersion

* chore(deps): bump updater version